@@ -13,16 +13,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.1.0",
+    "@angular/animations": "^14.2.6",
     "@angular/cdk": "^13.1.0",
-    "@angular/common": "^13.1.0",
-    "@angular/compiler": "^13.1.0",
-    "@angular/core": "^13.1.0",
-    "@angular/forms": "^13.1.0",
+    "@angular/common": "^14.2.6",
+    "@angular/compiler": "^14.2.6",
+    "@angular/core": "^14.2.6",
+    "@angular/forms": "^14.2.6",
     "@angular/material": "^13.1.0",
-    "@angular/platform-browser": "^13.1.0",
-    "@angular/platform-browser-dynamic": "^13.1.0",
-    "@angular/router": "^13.1.0",
+    "@angular/platform-browser": "^14.2.6",
+    "@angular/platform-browser-dynamic": "^14.2.6",
+    "@angular/router": "^14.2.6",
     "@asymmetrik/ngx-leaflet": "^8.1.0",
     "@asymmetrik/ngx-leaflet-draw": "^7.0.0",
     "@boring.devs/ngx-datatable": "^20.0.0",
@@ -49,15 +49,15 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.1.1",
+    "@angular-devkit/build-angular": "^14.2.6",
     "@angular-eslint/builder": "13.0.1",
     "@angular-eslint/eslint-plugin": "13.0.1",
     "@angular-eslint/eslint-plugin-template": "13.0.1",
     "@angular-eslint/schematics": "13.0.1",
     "@angular-eslint/template-parser": "13.0.1",
-    "@angular/cli": "^13.1.1",
-    "@angular/compiler-cli": "^13.1.0",
-    "@angular/language-service": "^13.1.0",
+    "@angular/cli": "^14.2.6",
+    "@angular/compiler-cli": "^14.2.6",
+    "@angular/language-service": "^14.2.6",
     "@types/echarts": "^4.9.15",
     "@types/jasminewd2": "^2.0.10",
     "@types/leaflet": "^1.7.6",
