@@ -2,7 +2,7 @@ import { Component, OnInit, OnDestroy } from "@angular/core";
 import { UserService } from "../../services/user.service";
 import { Subscription } from "rxjs";
 import { User } from "../../models/user";
-import { FormGroup, FormControl, Validators } from "@angular/forms";
+import { UntypedFormGroup, UntypedFormControl, Validators } from "@angular/forms";
 import { ActivatedRoute } from "@angular/router";
 import { MessageService } from "@core/services/message.service";
 
@@ -16,7 +16,7 @@ export class UserSettingsComponent implements OnInit, OnDestroy {
   user: User;
   id: number;
   editMode: boolean;
-  userForm: FormGroup;
+  userForm: UntypedFormGroup;
   hide = true;
 
   constructor(
@@ -32,9 +32,9 @@ export class UserSettingsComponent implements OnInit, OnDestroy {
 
   // set up form
   initForm(user): void {
-    this.userForm = new FormGroup({
-      firstName: new FormControl(user.firstName, Validators.required),
-      lastName: new FormControl(user.lastName, Validators.required),
+    this.userForm = new UntypedFormGroup({
+      firstName: new UntypedFormControl(user.firstName, Validators.required),
+      lastName: new UntypedFormControl(user.lastName, Validators.required),
     });
   }
 
