@@ -1,5 +1,5 @@
 import { Component, EventEmitter, Input, OnInit, Output } from "@angular/core";
-import { FormControl, FormGroup, Validators } from "@angular/forms";
+import { UntypedFormControl, UntypedFormGroup, Validators } from "@angular/forms";
 import { WidgetProperties } from "@features/widget/models/widget";
 import {
   WidgetDisplayOption,
@@ -32,11 +32,11 @@ export class WidgetEditInfoComponent implements OnInit {
   widgetTypes: WidgetType[];
   statTypes: any;
 
-  widgetForm = new FormGroup({
-    name: new FormControl("", Validators.required),
-    type: new FormControl("", Validators.required),
-    stat: new FormControl("latest", Validators.required), // default is raw data
-    displayType: new FormControl(""),
+  widgetForm = new UntypedFormGroup({
+    name: new UntypedFormControl("", Validators.required),
+    type: new UntypedFormControl("", Validators.required),
+    stat: new UntypedFormControl("latest", Validators.required), // default is raw data
+    displayType: new UntypedFormControl(""),
   });
 
   constructor(widgetConfigService: WidgetConfigService) {
