@@ -12,7 +12,7 @@ import { Threshold } from "@features/widget/models/threshold";
 import { Metric } from "@core/models/metric";
 import { WidgetConfigService } from "@features/widget/services/widget-config.service";
 import { Subscription } from "rxjs";
-import { FormArray, FormBuilder, FormGroup, Validators } from "@angular/forms";
+import { UntypedFormArray, UntypedFormBuilder, UntypedFormGroup, Validators } from "@angular/forms";
 import * as colormap from "colormap";
 import { WidgetProperties } from "@features/widget/models/widget";
 import {
@@ -41,7 +41,7 @@ export class WidgetEditOptionsComponent
   gradientOptions: any[];
   solidOptions: any[];
 
-  optionsForm: FormGroup = this.formBuilder.group({
+  optionsForm: UntypedFormGroup = this.formBuilder.group({
     thresholdArray: this.formBuilder.array([]),
     options: this.formBuilder.group({
       inRange: [null],
@@ -53,7 +53,7 @@ export class WidgetEditOptionsComponent
 
   constructor(
     widgetConfigService: WidgetConfigService,
-    private formBuilder: FormBuilder
+    private formBuilder: UntypedFormBuilder
   ) {
     this.gradientOptions = widgetConfigService.gradientOptions;
     this.solidOptions = widgetConfigService.solidOptions;
@@ -235,8 +235,8 @@ export class WidgetEditOptionsComponent
     this.subscriptions.unsubscribe();
   }
 
-  get thresholdArray(): FormArray {
-    return this.optionsForm.get("thresholdArray") as FormArray;
+  get thresholdArray(): UntypedFormArray {
+    return this.optionsForm.get("thresholdArray") as UntypedFormArray;
   }
 
   // Add a new threshold
